+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 ## 0.0.49
-- [common] Add CubeShellIterator
+- [common] Add BoxSurfaceIterator
 
 ## 0.0.48
 - [spigot] Fix player quit handling in `VisibilityListener`